bug fix registration message appears in registration page correctly,
extract registration servuce into a separate file
