+++ src/app/registration/index.js
@@ -12,63 +12,42 @@ export const RegistrationCtrlState ={
   }
 };
 
-export const RegistrationServices = [
-
- '$http', '$state',
-
- class RegistrationServices {
-  constructor ($http, $state, registeruser) {
-    this.$http = $http;
-    this.$state = $state;
-    this.register = registeruser;
-  }
-
-  registerUser (registerData) {
-    this.data = registerData;
-    const req = {
-      method: 'POST',
-      url: `/api/register`,
-      headers: {
-        'Content-Type': 'application/x-www-form-urlencoded'
-      },
-      data: `username=${registerData.username}&password=${registerData.password}&=`
-     };
-     console.log('req: ', req);
-     return this.$http(req).success(response => {
-        if(response.success === true){
-          this.$state.go('splash', {registrationMessage: response.registrationMessage});
-        }else{
-          this.$state.go('registration', {errorMessage: response.errorMessage});
-        }
-       return response;
-     });
-  }
-}];
-
 export const RegistrationCtrl = [
-  '$scope', '$state', '$stateParams', 'RegistrationServices', '$rootScope',
-
+  '$scope',
+  '$state',
+  '$stateParams',
+  'RegistrationServices',
+  '$rootScope',
+  '$timeout',
   class RegistrationCtrl {
-    constructor($scope, $state, $stateParams, RegistrationServices, $rootScope) {
+    constructor($scope, $state, $stateParams, RegistrationServices, $rootScope, $timeout) {
       this.registerData ={
         username: '',
         password: ''
       };
+      this.$state = $state;
+      this.timeD = false;
 
       $scope.userName = '';
       $scope.password = '';
       $scope.RegistrationServices = RegistrationServices;
-      $scope.registrationMessage = '';
 
       $scope.registerUser = () => {
         this.registerData.username = $scope.userName;
         this.registerData.password = $scope.password;
         RegistrationServices.registerUser(this.registerData)
           .success(response =>{
-            $rootScope.user = response.username;
-            $rootScope.visible = true;
-            $scope.registrationMessage = 'User successfully created';
-        });
+            if (response.success === true) {
+              $rootScope.user = response.username;
+              $rootScope.visible = true;
+              if (this.timeD) {
+                $timeout.cancel(timer)
+              }
+              this.timeD = $timeout(() => {
+                this.$state.go('splash');
+              }, 3000);
+            }
+        })
       };
       $scope.errorMessage = $stateParams.errorMessage;
     }
