+++ src/app/registration/registration.html
@@ -5,5 +5,5 @@
   <p>Please create a password</p>
   <input ng-model='password' type="password" name="Password">
   <button ng-click='registerUser()'>Register</button>
-  <p>{{ errorMessage }} {{ registrationMessage }}</p>
+  <p>{{ errorMessage }}</p>
 </div>
